tec: Move utils functions in seperated files
