+++ client/src/utils/array.js
@@ -7,15 +7,6 @@ function mapElementsById (elements, fk = 'id') {
   return byIds
 }
 
-function objectsToOptions (objects, valueKey, labelKey) {
-  return objects.map(obj => {
-    return {
-      value: obj[valueKey],
-      label: obj[labelKey],
-    }
-  })
-}
-
 function groupByFirstCharacter (array, attribute = null) {
   let groups = {}
 
@@ -33,6 +24,5 @@ function groupByFirstCharacter (array, attribute = null) {
 
 export {
   mapElementsById,
-  objectsToOptions,
   groupByFirstCharacter,
 }
