+++ client/src/utils/object.js
@@ -1,0 +1,12 @@
+function objectsToOptions (objects, valueKey, labelKey) {
+  return objects.map(obj => {
+    return {
+      value: obj[valueKey],
+      label: obj[labelKey],
+    }
+  })
+}
+
+export {
+  objectsToOptions,
+}
